requirements.txt
azure-functions==1.3.0
azure-kusto-data==1.0.3
azure-kusto-ingest==1.0.3
azure-storage-blob==12.5.0
azure-keyvault==4.1.0
json5==0.9.5
pandas==1.1.3
python-dateutil==2.8.1
pytz==2020.1
wheel